+++ Internship-4-OOP2/Application/DTOs/Users/CreateUserDto.cs
@@ -7,8 +7,8 @@
         public string Email { get; set; } = null!;
         public string AddressStreet { get; set; } = null!;
         public string AddressCity { get; set; } = null!;
-        public decimal GeoLat { get; set; }
-        public decimal GeoLng { get; set; }
+        public double GeoLat { get; set; }
+        public double GeoLng { get; set; }
         public string? Website { get; set; }
 
     }
