+++ Internship-4-OOP2/Application/DTOs/Users/UpdateUserDto.cs
@@ -6,7 +6,7 @@
         public string AddressStreet { get; set; } = null!;
         public string AddressCity { get; set; } = null!;
         public string ?Website { get; set; }
-        public decimal GeoLat { get; set; }
-        public decimal GeoLng { get; set; }
+        public double GeoLat { get; set; }
+        public double GeoLng { get; set; }
     }
 }
