+++ Internship-4-OOP2/Application/DTOs/Users/UserResponseDto.cs
@@ -9,8 +9,8 @@
         public string AdressStreet { get; internal set; } = null!;
         public string AdressCity { get; internal set; } = null!;
         public string? Website { get; set; }
-        public decimal GeoLat { get; set; }
-        public decimal GeoLng { get; set; }
+        public double GeoLat { get; set; }
+        public double GeoLng { get; set; }
         public bool IsActive { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
